streamlit_app/requirements.txt
boto3==1.20.43
Pillow==9.0.0
pydantic==1.9.0
streamlit==1.3.1
urlextract==1.5.0
